requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
boto3==1.26.153
botocore==1.29.153
cffi==1.15.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.17.1
Flask==1.1.2
fonttools==4.40.0
greenlet==2.0.2
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.14.0
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
nbformat==5.9.0
nest-asyncio==1.5.6
numpy==1.21.2
packaging==23.1
pandas==1.4.4
paramiko==3.2.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
plotly==5.9.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
pyzmq==25.1.0
s3transfer==0.6.1
scikit-learn==1.0.2
scipy==1.10.1
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.39
sshtunnel==0.4.0
stack-data==0.6.2
tenacity==8.2.2
threadpoolctl==3.1.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.0.3
zipp==3.15.0